refactor(save_conversation): Move conversation model in a specific module
